requirements.txt
google-generativeai>=0.1.0  # replace with actual package name if different
google-genai>=1.50.0
opencv-python
mediapipe
numpy
pandas
gradio
matplotlib
tqdm
scikit-learn
